requirements.txt
###### Requirements without Version ######`
numpy
pandas
imagecodecs

###### Requirements with Version Specifiers ######`
tifffile >= 2020.11.26
scikit-image >= 0.17.2
Pillow >= 8.0.1
opencv-python >= 4.4.0.46
xarray == 0.16.2
